refactor discoverfields into own componment
